Add check in Install-DotFiles that we have Administrator privileges
This is technically the wrong thing to check: we should be checking that
we have the SeCreateSymbolicLinkPrivilege. However, the default security
policy is to only grant this user right to the Administrators group, and
even if the default is changed to grant it to other groups, this won't
work on systems where UAC is enabled if the user is an Administrator.

So the only edge case I can think of where this check is inadequate is
for standard users on systems where the default user rights assignment
has been modified to grant them SeCreateSymbolicLinkPrivilege. That's a
pretty extreme edge case...