Always calculate the install path for components
We can't return early, skipping calculation of the install path, even
for components which are not available. That's because the detection
method may change as a result of subsequent custom metadata, which may
result in a previously unavailable component now being detected. If that
custom metadata uses the installation path of the global metadata, then
installation will fail as it hasn't been set.